nix flock fix breaking change in syssted thing
from the letter:

we would like to inform you about a breaking change in NixOS unstable, which
will result in VPS not starting. The change has been introduced by the following
commit in nixpkgs:

https://github.com/NixOS/nixpkgs/commit/a3e0698bf6fd676d96b89bdb4cd54c73ea502746

Option boot.systemdExecutable now requires absolute path to the systemd
executable, but because the configuration file for NixOS in our environment
uses relative path, the systemd executable will not be found when the VPS
is starting. The solution is to update the configuration file, which the VPS
template installs at /etc/nixos/vpsadminos.nix.

The required change can be seen here:

https://github.com/vpsfreecz/vpsadminos/commit/e8101cc63938585f5cf7b692d8dab234267b957f

The whole file can be downloaded from here:

https://github.com/vpsfreecz/vpsadminos/raw/staging/os/lib/nixos-container/vpsadminos.nix

Since this change will in time land in the next NixOS stable release, we
recommend to update the configuration at your earliest convenience.
The up-to-date vpsadminos.nix file will work also on NixOS 21.11.

Should you still encounter this issue and the VPS is not starting, you can use
the start menu in remote VPS console in vpsAdmin to start the system from
an older generation, see:

https://kb.vpsfree.org/manuals/vps/start_menu

Best regards,